Vishalka rne (#723)
* Enable host-side RNE rounding if on ATS/PVC

This commit moves the rounding routines to a new file (lp.hpp) that
performs host-side rounding. If the macro LIFT_EMBARGO is defined,
we use routines from src/comp/bf16/bf16.hpp to perform RNE rounding
using CPU intrinsics on the host-side. If the mcro os not defined,
we fall back to truncation.

* Enable (embargo-ed) device-side RNE rounding if on ATS/PVC

This commit moves the device-side rounding routines to a new file
(lp.h) that performs device-side rounding. If the macro LIFT_EMBARGO
is defined, we use ATS?PVC harwdare instructions via the intrinsics
in rne.h to perform RNE rounding. If the macro is not defined, we
fall back to truncation.

* Add SpirV for ring_allreduce.cl with bf16 <-> fp32 conversions from lp.h & rne.h