Don't define arg options as a separate policies (#759)
* Don't define arg options as a separate policies

Previously arg that required uncached memory allocation
was defined as a separate policy, but this approach is
not convenient when more policies/argument options are
added. To simplify this case specify argument options as
a separate template parameter that adds corresponding
query checks(e.g. is_uncached()) and use these checks
to choose the right overload.

Signed-off-by: Ilya Stepykin <ilya.stepykin@intel.com>