Fix crash in zeModuleDestroy on exit
Previously we had a crash due to zeModuleDestroy being called
multiple times on the same handle. The reason for that is that
we had several device storages containing device referencing
the same modules(different isntances, but same underlying L0 handle)

The duplicate device storage was unnecessary so we can just remove
it and as a result fix the crash for now. It's not a full fix for the
crash which requires refactoring our module storage(added a corresponding
todo), but it's enough for now.

Signed-off-by: Ilya Stepykin <ilya.stepykin@intel.com>