Fix LP functional tests (#898)
Fix setting of LP(BF16/FP16) cmake environment for functional tests.
Functional tests are built by Jenkins as standalone project and previously LP environment was not available in that mode and LP datatypes were not tested. But if tests are built as part of the parent CCL project then LP env was available - so developer just built project and didn't see that issue.
Fixed this by moving LP env into separate function in cmake helpers to reuse it in parent project and in functional tests.

Simplify logic with LP prologue/epilogue to handle the whole send/recv buffers independent on in-place/out-of-place, type of collectives. Otherwise different scenarios will complicate selection of right count, right offset inside buffers.

Add printing of current test config to have ability to validate really used test parameters in future.

Remove excessive out-of-place for FT bcast, in-place test is enough.