Set group size to control resource consumption (#969)
The resource consumption can be controlled by using zeKernelSetGroupSize() invoke.
The calculation of consumption for a proccess is based on setting groupSizeX, and it can not be greater,
than the limit of maxTotalGroupSize (the value is relate to the host).

Important:
Changing the group size , which gives the process more or less resource consumption, affects to the performance.
This patch increases the group size, it means that resource consumption is increased too to the maximum per one group.
(groups can be several one. see maxGroupCountX\Y\Z)
groupSizeX = size_of_send_buffer / vector_size. where vector size is 4, becasue of using float4 for kernels.
the bfloat's vec size is 1.

Signed-off-by: Nikita Gusev <nikita.gusev@intel.com>