Use c++17 for build with dpcpp compiler (#1062)
Update the version so we won't get errors coming
from the compiler's headers.
Right now the minimum required version is c++14
but it seems to be updated to c++17 soon, so
just use c++17. It's already supported as it's
used by the compiler by default(without -std flag).

The current implementation has a limitation:
in case of old cmake(version < 3.1) there is a
different mechanism to set c++ standard and
it requires more work in order not to break
everything else. It will be fixed later.