Propagate ptr offset during IPC exchange (#1158)
In case if 2 pointer belong to the same L0 allocation
and passed via IPC they're represented as differed
file descriptors but when the pointers are retrieved
by the other process, the retrieved pointers are the same
and represent the base of the L0 allocation.

To fix this issue, additionally propagate an offset for
each pointer and the other process gets the corrent pointers
using the offset and the retrieved base pointer.

When we close ipc pointers(zeMemCloseIpcHandle) that belongs
to the same allocation we might get an error. This error is not
meaningful(the function essentially has free()-like semantic),
so we can safely skip errors there in order to support the offsets.