Make deps_entry non-blocking (#1199)
Instead of using wait() for each event in start(), use test()
in entry's update(). This way the entry is not blocked until
the event is completed.

Also to unify the code, implement test() method of native_event
for SYCL, this way we can simply call dep.test() independently
of the event type.