Move unix sockets under /tmp and make them unique  (#1216)
Move unix sockets under /tmp to avoid file creation in a local
directory.
Also made the socket names unique to avoid issues
with parallel runs. The uniqueness is done by appending
I_MPI_HYDRA_UUID if available to the socket name. This is a
temporal solution as it works only with mpiexec, eventually
we should use process pid(but this requires us to exchange pid
information among processes via kvs which is not done yet)