Do submit_barrier to sync with user's queue (#1223)
In order to sync with user's queue in case of in-order one
do submit_barrier. This way out operations such as kernel won't
start before all the predeceasing operations on the user's queue
are complete (this was possible before because we're using a separate
out-of-order queue).
The behavior in case of out-of-order queue is unchanged.