Fix missed sched barriers (#1516)
These are barriers for schedule only. To prevent worker progress until what happened before the barrier is completed. Here I am fixing an issue from another PR where barriers were carried over in a similar way but not everywhere. So, this is a fix, but these barriers have always been there before. Sched barriers necessary to avoid possible data race, because we can start the next entry here before the previous entry is completed. It is not expected behavior, especially for a function that is a barrier.