Implement stub backend support (#1557)
Now we have a notion of a backend with a native being our default with L0
Add a new stub backend with a purpose to be a basis for other implementations(i.e. nccl backend).
The backend dispatching(based on env variable) is added for some parts of the code:

Event types: for the new backend added stub_event, which is always complete(as stub_comm does nothing)
Internal communicator: now we can choose between ccl_comm and sub_comm. stub_comm basically does nothing and simply returns an always completed stub_event. The collective methods of stub_comm simply print a message for test purpose. By providing a default implementation for some collectives we keep the number of required methods as low as possible, basically these are only void* overloads without a couple non-implemented methods.
KVS implementation: for stub backend we have a special kvs impl: stub_kvs_impl, it basically stores an id(could be set via env variable for test purpose) and puts it into kvs::address_type and then passes it further. The stub_kvs_impl is propagated to stub_comm, where the id can be retrieved. This approach allows us to have the same kvs handling for different backend(assuming the user uses our default kvs).
Also add a simple test to check the behavior: simply set an id via env variable and check the output string from collective. This way we verify that id is successfully propagated to the communicator and we call the right methods.