sched: enable cache for restarted operations (#1619)
Implements functionality necessary to enable caching of restarted operations.

Restarted operation means that operation was already cached
and user starts new one (i.e. with the same signature)
w/o waiting completion for previous one(s).

The basic idea is that now we detect if the schedule retrieved
from the cache is being executed. In this case we don't restart it/
start a new one, but instead create a new request (with all the
necessary events and the barrier), remember it and return to the user request,
to allow track the execution. Once the schedule completes,
worker thread checks whether there are requests for execution, if so,
restarts the schedule using the new request.

Introduced a new dynamic event pool to simplify allocation/deallocation
of level zero events necessary for caching functionality of restarted ops.
Now the pools are shared among scheds with the same context.

allreduce_events_inorder.cpp has been modified to test the new use-case:
now there are 2 modes: single buffer (use a single buffer for all
kernels/collectives) and multi buffer (use different buffers in scope
of an iteration). Also added possibility to enable cache via API
in the testcase.
Added command line arguments for the test to set different parameters
instead of using env vars.